feat: add vitest-fetch-mock for enhanced testing and setup configuration
