+++ package.json
@@ -54,6 +54,7 @@
     "tsx": "^4.19.2",
     "typescript": "^5.7.3",
     "vitest": "^3.0.5",
+    "vitest-fetch-mock": "^0.4.3",
     "vitest-testdirs": "^2.1.1"
   },
   "pnpm": {
