+++ test/__setup.ts
@@ -1,0 +1,6 @@
+import { vi } from "vitest";
+import createFetchMock from "vitest-fetch-mock";
+
+const fetchMocker = createFetchMock(vi);
+
+fetchMocker.enableMocks();
