+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    setupFiles: [
+      "./test/__setup.ts",
+    ],
+  },
+});
